feat: refactor http server package
